CSD-460 refactor: Modular structure
