+++ requirements.txt
@@ -1,5 +1,6 @@
 Flask==3.1.1
 Flask-SQLAlchemy==3.1.1
 SQLAlchemy==2.0.36
-PyMySQL==1.1.1
+PyMySQL[rsa]==1.1.1
+cryptography==43.0.1
 python-dotenv==1.0.1
